added driveWithGyroSystem method
The Competition Prayer
Oh, great REV Expansion hub up in the sky,
Thou who in your nature is so gracious and professional,
On this fateful competition day, we ask for thy help in blessing this robot.

We ask you that our Mecanum wheels strafe with the grace of a ballerina,
our motors spin with the strength of typhoons,
and that our gears will stay meshed tooth by tooth.

We ask you to ensure that our solder joints hold fast,
That our sensors can discern colors like a renaissance artists,
And tell us limits and telemetry so that our robot may run true.

And although we’ve already asked so much of you you’re gracious professional,
We ask that you guide our autonomous with your steady hand,
 ensure every controller button and trigger press guides the robot well,
and that our robot scores many points in all its endeavors.

In the name of the Great REV Expansion Hub we pray,
AMEN.